+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * A `MasteringDisplayColourVolume` property to the `HeifImage` class.
   * The mastering display color volume data is used for HDR rendering.
   * Requires LibHeif version 1.15.0 or later.
+* A `PixelAspectRatio` property to the `HeifImage` class.
+  * Requires LibHeif version 1.15.0 or later.
 
 ## Removed